Support Rack 2
Rack 2 has renamed the file for the CommonLogger class. Rather
than support both file names, we'll release a new major version of
this gem that drops support for Rack < 2.0. As this is the only
change in the gem, the upgrade path for the few consumers of this
should be very easy.